Add files via upload
This repository contains the front-end code for the login page of a vehicle parking system. The login page serves as the initial point of interaction for users accessing the system. It provides a simple and intuitive interface for users to log in using their email and password credentials.
The login page features a clean and visually appealing design. It incorporates a custom CSS styling, complemented by Bootstrap CSS framework, to create an attractive and responsive layout. The use of background images, colors, and typography enhances the overall user experience.
The page includes a header section that prominently displays the title of the vehicle parking system, creating a consistent branding element throughout the application. A navigation bar is implemented, offering convenient links to various sections of the system, such as "About Us," "Profile," "Book Parking," "Feedback," "Contact Us," and "Payment."
The login form itself is enclosed within a dedicated container. It consists of input fields for email and password, allowing users to enter their login credentials securely. A "Login" button triggers the submission of the form data to the server-side script, enabling authentication and access control.
If the user is logged in (checked through the isLoggedIn() function), the script redirects them to the booking page. Otherwise, an alert message informs the user to log in first.